Facial-Recognition/requirements.txt
numpy
Pillow
boto3

